TD-3023 limit the removal of output content to scorm and html resource during resource unpublish activity.

+++ LearningHub.Nhs.WebUI/Controllers/ResourceController.cs
@@ -413,7 +413,11 @@
 
             if (validationResult.IsValid)
             {
-                _ = Task.Run(async () => { await this.fileService.PurgeResourceFile(associatedFile, null); });
+                if (associatedFile.ScormDetails != null || associatedFile.HtmlDetails != null)
+                {
+                    _ = Task.Run(async () => { await this.fileService.PurgeResourceFile(associatedFile, null); });
+                }
+
                 if (viewModel.CatalogueNodeVersionId == 1)
                 {
                     return this.Redirect("/my-contributions/unpublished");
